+++ mopidy/__main__.py
@@ -11,14 +11,6 @@ from mopidy import commands, config as config_lib, ext
 from mopidy.internal import encoding, log, path, process, versioning
 from mopidy.internal.gi import Gst  # noqa: F401
 
-try:
-    # Make GLib's mainloop the event loop for python-dbus
-    import dbus.mainloop.glib
-    dbus.mainloop.glib.threads_init()
-    dbus.mainloop.glib.DBusGMainLoop(set_as_default=True)
-except ImportError:
-    pass
-
 
 logger = logging.getLogger(__name__)